Simplify binary operator handling.
I had four paths for binary operators:
    Constant op Constant
    Variable op Constant
    Variable op Variable
    Constant op Variable

The first case is untouched, being wholly handled by the 'types'
package's Constant type. The other three cases were essentially
the same, but with replicated logic.

The idea was that we would apply optimisations based on the
constant parts. This isn't going to happen in the near future,
so I've had the second and fourth case convert the constants to
variables and then proceed to the third case.