Extend string support: concatenation, len, (fixed) println.
To implement string concatenaton in a sane way, a runtime
library is required. To do this, I've also had to implement
support for unsafe.Pointer.

I intend to write as much of the runtime library in Go as
possible. To support the Go-based runtime library, there
will be various "intrinsics" written directly in LLVM
bitcode (e.g. memory management, threading).