Start implementing expression typechecking.
This is fairly experimental, but seems to be workable.

According to gcimporter, untyped constants are given a "Basic"
Type. We now do the same. There's no special "rune" kind, so the
Int kind refers to an untyped integer, while Int32 kind refers to
an untyped character (Int32 corresponds to types.Rune.Underlying).