Modify TypeMap.ToLLVM to store/lookup types by their string.
Some types.Type structs have slice fields, which are not comparable.
So keying the types map on the type interface was leading to multiply
mapped types, which would not be considered equal by LLVM.

There are two outstanding problems with the new approach:
 - No differentiation between types from different packages.
 - It's slow.

It may be worthwhile investigating implementing a custom hashmap later.