Handle constant binary expressions involving different types.
The expression type was being computed incorrectly in
typechecking, which is now fixed, and the type is now
propagated to the codegen'ing.

Also, array lengths may now be given as a named constant. If
I need to I'll add handle other expressions, but I'd rather
just wait for exp/types...