Update llgo-dist to build syscall package.
The syscall package is copied from GOROOT when llgo-dist is run.
In the future, only the static files will be copied, and the
dynamic bits will be generated (like with the Perl script, but
probably with a Go program).

llgo will provide .ll with inline assembly to implement
Syscall, Syscall6, etc.