Implement extension for specifying attributes on global variables.
If a package-level variable is preceded with a comment starting
with "#llgo", then that comment will be parsed for attributes that
will be applied to the global variable.

So far I've only implemented a "linkage" attribute for specifying
the linkage type of a global variable. Later it may be desirable to
introduce additional attributes, e.g. for changing symbol names,
e.g. to refer to "@llvm." intrinsics from Go code.