Add interface support for receivers wider than a word
When a receiver-function is being declared or defined now,
a companion function is defined if the receiver is wider
than a word. The companion function simply loads its first
argument, and then passes that and all remaining arguments
to the standard function, returning the results.

Also, reenabled some old interface tests that now pass.