llgo: Update *LLVMValue to use go/types
Got rid of ConstValue; go/types folds constant expressions
and evaluates their types, so we can convert directly to
LLVM constants. Introduced a simpler "NewConstValue" which
takes a constant value and type, and returns a new *LLVMValue.

Also got rid of NilValue. We can check if LLVM values are
constant nulls instead.

I intend to get rid of TypeValue as well; when this is done,
LLVMValue can be renamed to Value, and the interface can be
got rid of.