Construct method sets properly (sort of)
This simplifies and fixes a great deal of code, in
particular: (static) value-to-interface conversions,
method selectors, and runtime type construction.

There are still some problems with the way method sets
are constructed: infinite recursion through embedded
types is one, and ignoring of fields with the same name
as methods is another.