Sort method sets
Runtime types' methods are expected to be sorted,
and various bits of the compiler assume method
sets to be in name order.

It would be nice to have interfaces' methods
sorted when they come out of go/types, which would
speed things up a bit. I'll put together a CL
some time.