Generate program initialization code at compile time
Previously we were creating an "appending" linkage
array in runtime, containing a list of init functions.
This was causing difficult to debug issues when the
declared array is greater than size one (I think that's
the issue, anyway).

Instead, now, we just generate a function that calls
each of the (transitively) imported packages' init
functions. go.tools/ssa gurantees an init function for
each package, and we know the order up front, so this
is safe to do (and simplifies things somewhat).