cmd/llgo-build: record and use CGO_LDFLAGS
We now write out <package>.ldflags for each
cgo-using package built with llgo-build. When
building a command, llgo-build will check to
see if there is an .ldflags file and add the
flags within to the final link flags.

With these changes, it is now possible to
build llgo with llgo: the actual resulting
binary does not work correctly, though.