Don't parse runtime package
Parsing the runtime package was previously done
to avoid discrepancies between frontend and
runtime definitions of runtime functions and
types. This does carry a cost of tight coupling,
and increases the startup overhead of the
compiler.

We now define all runtime functions in the
compiler, and no longer parse the runtime
package.