Initial version of gllgo driver
This is intended to be a drop-in replacement for the gccgo executable that
can be used by the libgo build system, "go build" etc.

In the long term, we can consider working with the upstream Go authors to
adopt a subset of the gccgo command syntax as a common interface between
"go build" and a compiler.