Compute sizes and offsets ourselves with padding for embedded structures
The implementation is a tweaked version of the StdSizes implementation which
pads structures to align them. This is similar to what we were previously doing
in llvmTypeMap.Sizeof, but since StdSizes.Offsetsof calls StdSizes.Sizeof
we were getting it wrong when a struct which required padding was embedded
in another struct.